Fix migrate settings #309
